gui: fixes a forever loop when a malformed interface is specified
If you used an argument that results in the device failing to open, it falls through to Exception. If the exception occurs, you can result in a forever loop where the interface was set at the command line, but the device fails to open.

If an exception occurs, force a return to the setup window, with cleared iface and args.